# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MATOS CLAP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/items/": {
      "get": {
        "tags": [
          "items"
        ],
        "summary": "Get Items",
        "operationId": "items_get_items",
        "parameters": [
          {
            "name": "availability",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Availability"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Availability"
            }
          },
          {
            "name": "condition",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Condition"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Condition"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemPublic"
                  },
                  "title": "Response Items Get Items"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "items"
        ],
        "summary": "Create Item",
        "operationId": "items_create_item",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemPost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemPublic"
                }
              }
            }
          },
          "404": {
            "description": "Referenced catalog not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/items/loaned": {
      "get": {
        "tags": [
          "items"
        ],
        "summary": "Get Loaned Items",
        "description": "Get all items that are loaned out during a date range.\n\nReturns items with their associated loan information, useful for\nseeing what equipment is out at any given time.",
        "operationId": "items_get_loaned_items",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Start of the date range",
              "title": "Start Date"
            },
            "description": "Start of the date range"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "End of the date range",
              "title": "End Date"
            },
            "description": "End of the date range"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanedItemsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid date range"
          }
        }
      }
    },
    "/api/items/{item_id}": {
      "get": {
        "tags": [
          "items"
        ],
        "summary": "Get Item By Id",
        "operationId": "items_get_item_by_id",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemPublic"
                }
              }
            }
          },
          "404": {
            "description": "Item not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "items"
        ],
        "summary": "Update Item",
        "operationId": "items_update_item",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Item Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemPublic"
                }
              }
            }
          },
          "404": {
            "description": "Item or referenced catalog not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "items"
        ],
        "summary": "Soft Delete Item",
        "operationId": "items_soft_delete_item",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Item Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Item not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Availability": {
        "type": "string",
        "enum": [
          "available",
          "maintenance",
          "retired"
        ],
        "title": "Availability",
        "description": "Represents the availability status of an item."
      },
...
      "CatalogPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryPublic"
          },
          "image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Path"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "category"
        ],
        "title": "CatalogPublic"
      },
...
      "CategoryPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CategoryPublic"
      },
      "Condition": {
        "type": "string",
        "enum": [
          "new",
          "good",
          "degraded"
        ],
        "title": "Condition",
        "description": "Represents the condition of an item in the inventory."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemPatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "catalog_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Id"
          },
          "condition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Condition"
              },
              {
                "type": "null"
              }
            ]
          },
          "availability": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Availability"
              },
              {
                "type": "null"
              }
            ]
          },
          "deposit_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit Cents"
          }
        },
        "type": "object",
        "title": "ItemPatch"
      },
      "ItemPost": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "catalog_id": {
            "type": "integer",
            "title": "Catalog Id"
          },
          "condition": {
            "$ref": "#/components/schemas/Condition"
          },
          "availability": {
            "$ref": "#/components/schemas/Availability",
            "default": "available"
          },
          "deposit_cents": {
            "type": "integer",
            "title": "Deposit Cents",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name",
          "catalog_id",
          "condition"
        ],
        "title": "ItemPost"
      },
      "ItemPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "catalog": {
            "$ref": "#/components/schemas/CatalogPublic"
          },
          "condition": {
            "$ref": "#/components/schemas/Condition"
          },
          "availability": {
            "$ref": "#/components/schemas/Availability"
          },
          "deposit_cents": {
            "type": "integer",
            "title": "Deposit Cents"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "catalog",
          "condition",
          "availability",
          "deposit_cents"
        ],
        "title": "ItemPublic"
      },
...
      "LoanStatus": {
        "type": "string",
        "enum": [
          "scheduled",
          "active",
          "returned"
        ],
        "title": "LoanStatus",
        "description": "Status of a loan for timeline display."
      },
...
      "LoanedItemWithLoan": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "item_name": {
            "type": "string",
            "title": "Item Name"
          },
          "catalog_name": {
            "type": "string",
            "title": "Catalog Name"
          },
          "loan_id": {
            "type": "integer",
            "title": "Loan Id"
          },
          "borrower_name": {
            "type": "string",
            "title": "Borrower Name"
          },
          "loan_start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Loan Start Date"
          },
          "loan_end_date": {
            "type": "string",
            "format": "date-time",
            "title": "Loan End Date"
          },
          "actual_start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Start Date"
          },
          "actual_return_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Return Date"
          },
          "status": {
            "$ref": "#/components/schemas/LoanStatus"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "item_name",
          "catalog_name",
          "loan_id",
          "borrower_name",
          "loan_start_date",
          "loan_end_date",
          "status"
        ],
        "title": "LoanedItemWithLoan",
        "description": "An item that is currently loaned out, with loan details."
      },
      "LoanedItemsResponse": {
        "properties": {
          "start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date-time",
            "title": "End Date"
          },
          "loaned_items": {
            "items": {
              "$ref": "#/components/schemas/LoanedItemWithLoan"
            },
            "type": "array",
            "title": "Loaned Items"
          }
        },
        "type": "object",
        "required": [
          "start_date",
          "end_date",
          "loaned_items"
        ],
        "title": "LoanedItemsResponse",
        "description": "Response for querying items that are loaned out during a date range."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}